refactor: 컴포넌트 폴더 분리 고도화
