+++ src/components/container/signUp/BackgroundPointContainer.tsx
