+++ src/components/items/MainBoardItem.tsx
