+++ src/routes/board/NormalBoard.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { userInfoState } from '../../atoms';
-import BoardHeader from '../../components/board/BoardHeader';
-import PostList from '../../components/board/PostList';
-import BoardListFooter from '../../components/board/BoardListFooter';
-import NavContainer from '../../components/board/NavContainer';
+import BoardHeader from '../../components/container/board/BoardHeader';
+import PostList from '../../components/container/board/PostList';
+import BoardListFooter from '../../components/container/board/BoardListFooter';
+import NavContainer from '../../components/container/board/NavContainer';
 import { BoardURLParams, Post } from '../../types/board';
 import { fetchPostList } from '../../api/board';
